db Änderungen, Änderung des Schemata, Implementation Controller,Service,Repo

+++ backend/EventTom/src/main/java/API/EventTom/services/interfaces/IEventService.java
@@ -9,6 +9,6 @@ public interface IEventService {
 
 
     public List<EventDTO> getAllEvents();
-    public EventDTO getEventById(int id);
+    public EventDTO getEventById(long id);
 
 }
